refactor(calendar): extract utility hooks and streamline imports
