+++ packages/components/calendar/src/index.ts
@@ -1,7 +1,16 @@
 export { Calendar } from "./calendar"
 export type { CalendarProps } from "./calendar"
 export {
+  getFirstOfWeek,
+  getLastOfWeek,
+  getWeekdays,
+  getMonthDays,
+  getRangeYears,
+  getRangeMonths,
+  getRangeDates,
+  getFormattedLabel,
   sortDates,
+  isInRange,
   isIncludeDates,
   isMonthInRange,
   isSameMonth,
@@ -12,7 +21,7 @@ export {
   isBeforeDate,
   isAfterMonth,
   isBeforeMonth,
-} from "./use-calendar"
+} from "./calendar-utils"
 export { DatePicker } from "./date-picker"
 export type { DatePickerProps } from "./date-picker"
 export { MonthPicker } from "./month-picker"
