+++ packages/components/calendar/src/year-list.tsx
@@ -6,11 +6,9 @@ import { cx, filterUndefined } from "@yamada-ui/utils"
 import type { FC } from "react"
 import type { CalendarHeaderProps } from "./calendar-header"
 import { CalendarHeader } from "./calendar-header"
-import {
-  getFormattedLabel,
-  useCalendarContext,
-  useYearList,
-} from "./use-calendar"
+import { getFormattedLabel } from "./calendar-utils"
+import { useCalendarContext } from "./use-calendar"
+import { useYearList } from "./use-year-list"
 
 type YearListOptions = {
   /**
